+++ public/components/add-modules-data/guides/aws.js
@@ -9,13 +9,15 @@
 *
 * Find more information about this on the LICENSE file.
 */
+import { webDocumentationLink } from "../../../../common/services/web_documentation";
+
 export default {
   id: 'aws',
   name: 'Amazon AWS services',
   wodle_name: 'aws-s3',
   description: 'Configuration options of the AWS-S3 wodle.',
   category: 'Security information management',
-  documentation_link: 'https://documentation.wazuh.com/current/user-manual/reference/ossec-conf/wodle-s3.html',
+  documentation_link: webDocumentationLink('user-manual/reference/ossec-conf/wodle-s3.html'),
   icon: 'logoAWSMono',
   avaliable_for_manager: true,
   avaliable_for_agent: true,