+++ server/start/initialize/index.ts
@@ -15,11 +15,10 @@ import { pluginPlatformTemplate } from '../../integration-files/kibana-template'
 import { getConfiguration } from '../../lib/get-configuration';
 import { totalmem } from 'os';
 import fs from 'fs';
-import { ManageHosts } from '../../lib/manage-hosts';
-import { WAZUH_ALERTS_PATTERN, WAZUH_DATA_CONFIG_REGISTRY_PATH, WAZUH_PLUGIN_PLATFORM_TEMPLATE_NAME, WAZUH_DATA_PLUGIN_PLATFORM_BASE_ABSOLUTE_PATH, PLUGIN_PLATFORM_NAME, PLUGIN_PLATFORM_INSTALLATION_USER_GROUP, PLUGIN_PLATFORM_INSTALLATION_USER } from '../../../common/constants';
+import { WAZUH_ALERTS_PATTERN, WAZUH_DATA_CONFIG_REGISTRY_PATH, WAZUH_PLUGIN_PLATFORM_TEMPLATE_NAME, WAZUH_DATA_PLUGIN_PLATFORM_BASE_ABSOLUTE_PATH, PLUGIN_PLATFORM_NAME, PLUGIN_PLATFORM_INSTALLATION_USER_GROUP, PLUGIN_PLATFORM_INSTALLATION_USER, WAZUH_DEFAULT_APP_CONFIG } from '../../../common/constants';
 import { createDataDirectoryIfNotExists } from '../../lib/filesystem';
+import _ from 'lodash';
 
-const manageHosts = new ManageHosts();
 
 export function jobInitializeRun(context) {
   const PLUGIN_PLATFORM_INDEX = context.server.config.kibana.index;
@@ -36,11 +35,6 @@ export function jobInitializeRun(context) {
       configurationFile && typeof configurationFile.pattern !== 'undefined'
         ? configurationFile.pattern
         : WAZUH_ALERTS_PATTERN;
-    // global.XPACK_RBAC_ENABLED =
-    //   configurationFile &&
-    //     typeof configurationFile['xpack.rbac.enabled'] !== 'undefined'
-    //     ? configurationFile['xpack.rbac.enabled']
-    //     : true;
   } catch (error) {
     log('initialize', error.message || error);
     context.wazuh.logger.error(
@@ -60,7 +54,7 @@ export function jobInitializeRun(context) {
   }
 
   // Save Wazuh App setup
-  const saveConfiguration = async () => {
+  const saveConfiguration = async (hosts = {}) => {
     try {
       const commonDate = new Date().toISOString();
 
@@ -70,15 +64,20 @@ export function jobInitializeRun(context) {
         revision: packageJSON.revision,
         installationDate: commonDate,
         lastRestart: commonDate,
-        hosts: {}
+        hosts
       };
       try {
         createDataDirectoryIfNotExists();
         createDataDirectoryIfNotExists('config');
+        log(
+          'initialize:saveConfiguration',
+          `Saving configuration in registry file: ${JSON.stringify(configuration)}`,
+          'debug'
+        );
         await fs.writeFileSync(WAZUH_DATA_CONFIG_REGISTRY_PATH, JSON.stringify(configuration), 'utf8');
         log(
           'initialize:saveConfiguration',
-          'Wazuh configuration registry inserted',
+          'Wazuh configuration registry saved.',
           'debug'
         );
       } catch (error) {
@@ -90,59 +89,94 @@ export function jobInitializeRun(context) {
     } catch (error) {
       log('initialize:saveConfiguration', error.message || error);
       context.wazuh.logger.error(
-        'Error creating wazuh-version registry'
+        'Error creating wazuh-registry.json file.'
       );
     }
   };
 
   /**
-   * Checks if the .wazuh-version exists, in this case it will be deleted and the wazuh-registry.json will be created
+   * Checks if the .wazuh-registry.json file exists:
+   * - yes: check the plugin version and revision match the values stored in the registry file.
+   *  If not, then it migrates the data rebuilding the registry file.
+   * - no: create the file with empty hosts
    */
   const checkWazuhRegistry = async () => {
-    try {
+    log(
+      'initialize:checkwazuhRegistry',
+      'Checking wazuh-registry.json file.',
+      'debug'
+    );
+
+    if(!fs.existsSync(WAZUH_DATA_PLUGIN_PLATFORM_BASE_ABSOLUTE_PATH)){
+      throw new Error(`The data directory is missing in the ${PLUGIN_PLATFORM_NAME} root instalation. Create the directory in ${WAZUH_DATA_PLUGIN_PLATFORM_BASE_ABSOLUTE_PATH} and give it the required permissions (sudo mkdir ${WAZUH_DATA_PLUGIN_PLATFORM_BASE_ABSOLUTE_PATH};sudo chown -R ${PLUGIN_PLATFORM_INSTALLATION_USER}:${PLUGIN_PLATFORM_INSTALLATION_USER_GROUP} ${WAZUH_DATA_PLUGIN_PLATFORM_BASE_ABSOLUTE_PATH}). After restart the ${PLUGIN_PLATFORM_NAME} service.`);
+    };
+
+    if (!fs.existsSync(WAZUH_DATA_CONFIG_REGISTRY_PATH)) {
       log(
         'initialize:checkwazuhRegistry',
-        'Checking wazuh-version registry.',
+        'wazuh-registry.json file does not exist. Initializing configuration.',
         'debug'
       );
 
-      if(!fs.existsSync(WAZUH_DATA_PLUGIN_PLATFORM_BASE_ABSOLUTE_PATH)){
-        throw new Error(`The data directory is missing in the ${PLUGIN_PLATFORM_NAME} root instalation. Create the directory in ${WAZUH_DATA_PLUGIN_PLATFORM_BASE_ABSOLUTE_PATH} and give it the required permissions (sudo mkdir ${WAZUH_DATA_PLUGIN_PLATFORM_BASE_ABSOLUTE_PATH};sudo chown -R ${PLUGIN_PLATFORM_INSTALLATION_USER}:${PLUGIN_PLATFORM_INSTALLATION_USER_GROUP} ${WAZUH_DATA_PLUGIN_PLATFORM_BASE_ABSOLUTE_PATH}). After restart the ${PLUGIN_PLATFORM_NAME} service.`);
-      };
+      // Create the app registry file for the very first time
+      await saveConfiguration();
+    } else {
+      // If this function fails, it throws an exception
+      const source = JSON.parse(fs.readFileSync(WAZUH_DATA_CONFIG_REGISTRY_PATH, 'utf8'));
 
-      if (!fs.existsSync(WAZUH_DATA_CONFIG_REGISTRY_PATH)) {
+      // Check if the stored revision differs from the package.json revision
+      const isUpgradedApp = packageJSON.revision !== source.revision || packageJSON.version !== source['app-version'];
+
+      // Rebuild the registry file if revision or version fields are differents
+      if (isUpgradedApp) { 
         log(
           'initialize:checkwazuhRegistry',
-          'wazuh-version registry does not exist. Initializing configuration.',
-          'debug'
+          'Wazuh app revision or version changed, regenerating wazuh-registry.json.',
+          'info'
         );
 
-        // Create the app registry file for the very first time
-        await saveConfiguration();
-      } else {
-        // If this function fails, it throws an exception
-        const source = JSON.parse(fs.readFileSync(WAZUH_DATA_CONFIG_REGISTRY_PATH, 'utf8'));
+        // Rebuild the registry file `wazuh-registry.json`
 
-        // Check if the stored revision differs from the package.json revision
-        const isUpgradedApp = packageJSON.revision !== source.revision || packageJSON.version !== source['app-version'];
+        // Get the supported extensions for the installed plugin
+        const supportedDefaultExtensionsConfiguration = Object.entries(WAZUH_DEFAULT_APP_CONFIG)
+          .filter(([setting]) => setting.startsWith('extensions.'))
+          .map(([setting, settingValue]) => {
+            return [setting.split('.')[1], settingValue];
+        });
 
-        // Rebuild the registry file if revision or version fields are differents
-        if (isUpgradedApp) { 
-          log(
-            'initialize:checkwazuhRegistry',
-            'Wazuh app revision or version changed, regenerating wazuh-version registry',
-            'info'
-          );
-          // Rebuild registry file in blank
-          await saveConfiguration();
-        }
+        // Get the supported extensions by ID
+        const supportedDefaultExtensionsNames = supportedDefaultExtensionsConfiguration.map(([setting]) => setting);
+
+        // Generate the hosts data, migrating the extensions.
+        // Keep the supported and existent extensions for the installed plugin with the configurated value
+        // Add the extensions with default values that didn't exist in the previous configuration
+        // Remove the unsupported extensions for the installed plugin
+        const registryHostsData = Object.entries(source.hosts).reduce((accum, [hostID, hostData]) => {
+          accum[hostID] = hostData;
+          if(accum[hostID].extensions){
+            // Migrate extensions to those supported by the installed plugin
+            const defaultHostExtentionsConfiguration = Object.fromEntries(supportedDefaultExtensionsConfiguration);
+            // Select of current configuration the extension IDs that are supported in the installed plugin
+            const currentHostConfiguration = _.pick(accum[hostID].extensions, supportedDefaultExtensionsNames);
+            // Merge the default extensions configuration with the configuration stored in the registry file
+            accum[hostID].extensions = _.merge(defaultHostExtentionsConfiguration, currentHostConfiguration);
+          }
+          return accum;
+        }, {});
+
+        // Rebuild the registry file with the migrated host data (extensions are migrated to these supported by the installed plugin).
+        await saveConfiguration(registryHostsData);
+
+        log(
+          'initialize:checkwazuhRegistry',
+          'Migrated the registry file.',
+          'info'
+        );
       }
-    } catch (error) {
-      return Promise.reject(error);
     }
   };
 
-  // Init function. Check for "wazuh-version" document existance.
+  // Init function. Check for wazuh-registry.json file exists.
   const init = async () => {
     await checkWazuhRegistry();
   };
@@ -187,7 +221,6 @@ export function jobInitializeRun(context) {
         'debug'
       );
       await init();
-      return;
     } catch (error) {
       return Promise.reject(
         new Error(
@@ -208,7 +241,6 @@ export function jobInitializeRun(context) {
         'debug'
       );
       await createEmptyKibanaIndex();
-      return;
     } catch (error) {
       return Promise.reject(
         new Error(
@@ -231,7 +263,6 @@ export function jobInitializeRun(context) {
         'debug'
       );
       await createEmptyKibanaIndex();
-      return;
     } catch (error) {
       log('initialize:getTemplateByName', error.message || error);
       return fixKibanaTemplate();